+++ contracts/compliance/modular/modules/CountryAllowModule.sol
@@ -67,8 +67,6 @@ import "../../../token/IToken.sol";
 import "./AbstractModule.sol";
 
 contract CountryAllowModule is AbstractModule {
-    string constant public NAME = "CountryAllowModule";
-
     /// Mapping between country and their allowance status per compliance contract
     mapping(address => mapping(uint16 => bool)) private _allowedCountries;
 
@@ -194,6 +192,10 @@ contract CountryAllowModule is AbstractModule {
      */
     function isCountryAllowed(address _compliance, uint16 _country) public view returns (bool) {
         return _allowedCountries[_compliance][_country];
+    }
+
+    function name() public pure returns (string memory _name) {
+        return "CountryAllowModule";
     }
 
     /**