+++ contracts/compliance/modular/modules/CountryRestrictModule.sol
@@ -67,8 +67,6 @@ import "../../../token/IToken.sol";
 import "./AbstractModule.sol";
 
 contract CountryRestrictModule is AbstractModule {
-    string constant public NAME = "CountryRestrictModule";
-
     /// Mapping between country and their restriction status per compliance contract
     mapping(address => mapping(uint16 => bool)) private _restrictedCountries;
 
@@ -196,6 +194,10 @@ contract CountryRestrictModule is AbstractModule {
     function isCountryRestricted(address _compliance, uint16 _country) public view
     returns (bool) {
         return ((_restrictedCountries[_compliance])[_country]);
+    }
+
+    function name() public pure returns (string memory _name) {
+        return "CountryRestrictModule";
     }
 
     /**